Add pt CLI + pt device refactor
