+++ .deprecated/ptoled/__init__.py
@@ -1,14 +1,13 @@
 from pitop.miniscreen.oled import (  # noqa: F401
     canvas,
     oled_image,
-    oled_manager,
 )
 from pitop.miniscreen.oled.core import (  # noqa: F401
     display,
     fps_regulator,
 )
 from pitop.miniscreen.oled import OLEDDisplay as PTOLEDDisplay  # noqa: F401
-from pitop.miniscreen.oled import oled_manager as device_helper  # noqa: F401
+from pitop.miniscreen.oled import oled_controls as device_helper  # noqa: F401
 
 print("Note: Use of the 'ptoled' package is now deprecated. Please use 'pitop.miniscreen.oled' instead.")
 # TODO
