+++ pitop/miniscreen/oled/__init__.py
@@ -1,6 +1,5 @@
 from .oled_display import OLEDDisplay  # noqa: F401
 from .oled_image import OLEDImage  # noqa: F401
-# TODO: make internal; expose "get/set control" and "get OLED object"
 from .oled_controls import (  # noqa: F401
     get_device_instance,
     device_reserved,
